[#1554] V2 namespace fix for feedback tryout
There is no V1 feedback model so this was a no-brainer fix to make. It
opens a can of jorms though which is related to how the tryouts tests
are currently organized (i.e. no separate dirs for v1, v2). We've
resolved this on a forward branch which I think I'll backport into
develop at some point.