[#1600] Update expectations for secrets now setting key in init
Previously we lazyily waited to generate a secret/metadata key until
just before saving. With a recent update to familia::horreum which now
has a more conventionally minded class-level create method, we need
the key generated earlier in order for save_if_not_exists to work
properly (otherwise it tries to check if exists but raises an error
when it can't put the dbkey together without the unique identifier.
The create change mattered more for models with natural keys like
email address or username but it also works just fine for models with
unique identifiers.