Skip some test coverage added from orphan branch
Not all of the tests we brought over have ever been passing. The coverage
area is good but they were written prior to the Rack::Session changes.
We may end up removing them and/or rewriting them, but keeping for now
as part of this PR. They're easy enough to be removed later if we
decide not to (they use the same skip message).