[#1726] Important: model db config should stay through v1.0
It's a non-issue for new installs, which logical database index models
save their records to. For existing installs that are updating (likely
via :latest OCI tag), changing the DB will functionally appear through
the UI as though all data is gone. It requires data migrations to
accomplish properly.

So at load time right after connecting to the DB, we now set a flag
for each model it its chosen DB to indicate that a data migration will
be needed for that logical DB. We'll use this at or around a v1.0
release when something as disruptive as a data migration is both
tolerable and not unexpected.