[#OCT18] Fix for missing index updates on Customer.create!
The Horreum.create! method uses save_if_not_exists as a precaution
which for some reason does not auto-update all unique indexes like
regular save does; it's only updating email. This caused an issue
when auth mode=advanced where when the customer record is being
created the extid index was not populated so on successful
authentication the customer record is not loaded from valkey
(the external ID field is the common link between databases).

We didn't see this behaviour when auth mode=basic b/c that goes
through V2::Logic::CreateAccount which makes an additional save
after create! which at that point populated the extid and objid
indexes.

This is likely a bug upstream in familia (need to do more digging)
but the fix for now is to simply call save from Rodauth's hook
after_create_account. It would be better for it to use the same
code path via the logic class. This is working now so I wanted
to get it checked in with an explanation as a known working version.