[i18n] Further fixes related to structural differences
As a follow-up to the previous commit (0308237cd9), it wasn't just the
one "dashboard" key that changed type. There were others ("welcome",
"stats", "domains", ...) so these changes were produced by the new
fix-locale-conflicts script.

I'll follow up again with running the harmonize script on them as well
since we have a higher trust level in the results it produces.