[#I18N] Fix structural conflicts
e.g. where en has key->[string] and the locale has key->[object] or
verse vicea.

Process:
  $ src/locales/scripts/harmonize/fix-all-locale-conflicts
  $ src/locales/scripts/harmonize/harmonize-all-locale-files

N.B. Running the fix script (which is based on the harmonize one)
twice produces the same results. It's possible that it supersedes
the harmonize script now since it can handle the structural
differences as well as. The second run was pretty simple though, just
removing a duplicate set of keys. The magic of harmonize is that it
compares the order of the keys as well so that the files share an
identical literal structure and not just conceptually. I need more
time see how the fix script handles those scenarios so we'll keep both
around for a little while longer.