[#2111] Fix event limit lookup
When register_events is called at start time, it is passed the
rate_limit settings as-is. With the config change from symbol keys to
strings, the matching of rate limit event to the hash of pre-configured
events didn't match. The result was all events regardless of
configuration were being limited by the default (25). Now events are
treated as strings as well so that they match the config setting
correctly.