Add an empty option for manual runs
Starting a build with the first element in the Target platform dropdown
selected ("linux/amd64,linux/arm64") produced a build with Run
PLATFORMS="linux/amd64". I'm wondering if the comma is an issue as an
issue in the dropdown component of the UI. Trying it with an empty
option to see if it will then grab the default instead.